Fix names of some reference functions
The `__name__` field of some binary reference functions was wrong. We
fix this to be consistent with unary reference functions. In the future,
we should probably make the binary reference wrapper return a wrapper
itself to avoid all those calls to `partial`.

This change helps performing some homogeneous treatment of functions by
their name.

ghstack-source-id: 7e7c5224e8fd3c5aab4ea75aa6f6fe2904d5a3b0
Pull Request resolved: https://github.com/pytorch/pytorch/pull/88115