extract out the clean workspace test to its own file
Summary:
This test relies on what the root workspace is before any other code
is run. However, some of the test cases change it. If the order the
tests are run is randomized, then the test can fail if run after one
of them.

Having it on its own ensures that it always sees a pristine state.

Test Plan:
Verified locally and confirmed in internal and external CI.

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 08f0824e29bf1d2017211668df0c8e4703b0fc29
Pull Request resolved: https://github.com/pytorch/pytorch/pull/88682